Add `Integer` type annotation to `sizehint!` overloads (#155)
JuliaLang/julia#59168 made `sizehint!(::ConcreteType, n)` use
`::Integer` instead of `::Any` for the second (`n`) argument. This
gives OrderedCollections the same behavior.

+++ src/ordered_dict.jl
@@ -212,7 +212,7 @@ function rehash!(h::OrderedDict{K,V}, newsz::Integer = length(h.slots)) where {K
     return h
 end
 
-function sizehint!(d::OrderedDict, newsz)
+function sizehint!(d::OrderedDict, newsz::Integer)
     slotsz = (newsz*3)>>1
     oldsz = length(d.slots)
     if slotsz <= oldsz
